Update matrix_mpi.h, replace everything with floats

+++ lib/matrix/matrix_mpi.h
@@ -43,13 +43,13 @@ void dot_mpi(Matrix mat1, Matrix mat2, Matrix product, int rank, int num_procs) 
     }
 
     // Scatter matrices mat1 and mat2
-    double *local_mat1 = malloc(send_counts[rank] * sizeof(double));
-    double *local_mat2 = malloc(mat2.rows * mat2.cols * sizeof(double));
-    MPI_Scatterv(mat1.data, send_counts, displs, MPI_DOUBLE, local_mat1, send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
-    MPI_Bcast(mat2.data, mat2.rows * mat2.cols, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    float *local_mat1 = malloc(send_counts[rank] * sizeof(float));
+    float *local_mat2 = malloc(mat2.rows * mat2.cols * sizeof(float));
+    MPI_Scatterv(mat1.data, send_counts, displs, MPI_FLOAT, local_mat1, send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
+    MPI_Bcast(mat2.data, mat2.rows * mat2.cols, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     // Compute local dot product
-    double *local_product = malloc(send_counts[rank] * mat2.cols * sizeof(double));
+    float *local_product = malloc(send_counts[rank] * mat2.cols * sizeof(float));
     int local_rows = send_counts[rank] / mat1.cols;
     for (int i = 0; i < local_rows; i++) {
         for (int j = 0; j < mat2.cols; j++) {
@@ -57,12 +57,12 @@ void dot_mpi(Matrix mat1, Matrix mat2, Matrix product, int rank, int num_procs) 
             for (int k = 0; k < mat1.cols; k++) {
                 sum += local_mat1[i * mat1.cols + k] * mat2.data[k * mat2.cols + j];
             }
-            local_product[i * mat2.cols + j] = sum;
+            local_product[i * mat2.cols + j] = (float)sum;
         }
     }
 
     // Gather local products to the master process
-    MPI_Gatherv(local_product, send_counts[rank], MPI_DOUBLE, product.data, send_counts, displs, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(local_product, send_counts[rank], MPI_FLOAT, product.data, send_counts, displs, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     free(local_mat1);
     free(local_mat2);
@@ -87,9 +87,9 @@ void add_mpi(Matrix mat1, Matrix mat2, int rank, int num_procs) {
     }
 
     // Scatter columns of mat1 and broadcast mat2
-    double *local_mat1 = malloc(mat1.rows * send_counts[rank] * sizeof(double));
-    MPI_Scatterv(mat1.data, send_counts, displs, MPI_DOUBLE, local_mat1, mat1.rows * send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
-    MPI_Bcast(mat2.data, mat2.rows * mat2.cols, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    float *local_mat1 = malloc(mat1.rows * send_counts[rank] * sizeof(float));
+    MPI_Scatterv(mat1.data, send_counts, displs, MPI_FLOAT, local_mat1, mat1.rows * send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
+    MPI_Bcast(mat2.data, mat2.rows * mat2.cols, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     // Add mat2 to local columns of mat1
     for (int i = 0; i < mat1.rows; i++) {
@@ -100,7 +100,7 @@ void add_mpi(Matrix mat1, Matrix mat2, int rank, int num_procs) {
     }
 
     // Gather the updated columns to the master process
-    MPI_Gatherv(local_mat1, mat1.rows * send_counts[rank], MPI_DOUBLE, mat1.data, send_counts, displs, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(local_mat1, mat1.rows * send_counts[rank], MPI_FLOAT, mat1.data, send_counts, displs, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     free(local_mat1);
     free(send_counts);
@@ -123,11 +123,11 @@ void subtract_mpi(Matrix mat1, Matrix mat2, Matrix difference, int rank, int num
     }
 
     // Scatter elements of mat1 and mat2
-    double *local_mat1 = malloc(send_counts[rank] * sizeof(double));
-    double *local_mat2 = malloc(send_counts[rank] * sizeof(double));
-    double *local_difference = malloc(send_counts[rank] * sizeof(double));
-    MPI_Scatterv(mat1.data, send_counts, displs, MPI_DOUBLE, local_mat1, send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
-    MPI_Scatterv(mat2.data, send_counts, displs, MPI_DOUBLE, local_mat2, send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    float *local_mat1 = malloc(send_counts[rank] * sizeof(float));
+    float *local_mat2 = malloc(send_counts[rank] * sizeof(float));
+    float *local_difference = malloc(send_counts[rank] * sizeof(float));
+    MPI_Scatterv(mat1.data, send_counts, displs, MPI_FLOAT, local_mat1, send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
+    MPI_Scatterv(mat2.data, send_counts, displs, MPI_FLOAT, local_mat2, send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     // Calculate local differences
     for (int i = 0; i < send_counts[rank]; i++) {
@@ -135,7 +135,7 @@ void subtract_mpi(Matrix mat1, Matrix mat2, Matrix difference, int rank, int num
     }
 
     // Gather the local differences to the master process
-    MPI_Gatherv(local_difference, send_counts[rank], MPI_DOUBLE, difference.data, send_counts, displs, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(local_difference, send_counts[rank], MPI_FLOAT, difference.data, send_counts, displs, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     free(local_mat1);
     free(local_mat2);
@@ -160,11 +160,11 @@ void hadamard_mpi(Matrix mat1, Matrix mat2, Matrix product, int rank, int num_pr
     }
 
     // Scatter elements of mat1 and mat2
-    double *local_mat1 = malloc(send_counts[rank] * sizeof(double));
-    double *local_mat2 = malloc(send_counts[rank] * sizeof(double));
-    double *local_product = malloc(send_counts[rank] * sizeof(double));
-    MPI_Scatterv(mat1.data, send_counts, displs, MPI_DOUBLE, local_mat1, send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
-    MPI_Scatterv(mat2.data, send_counts, displs, MPI_DOUBLE, local_mat2, send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    float *local_mat1 = malloc(send_counts[rank] * sizeof(float));
+    float *local_mat2 = malloc(send_counts[rank] * sizeof(float));
+    float *local_product = malloc(send_counts[rank] * sizeof(float));
+    MPI_Scatterv(mat1.data, send_counts, displs, MPI_FLOAT, local_mat1, send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
+    MPI_Scatterv(mat2.data, send_counts, displs, MPI_FLOAT, local_mat2, send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     // Calculate local product
     for (int i = 0; i < send_counts[rank]; i++) {
@@ -172,7 +172,7 @@ void hadamard_mpi(Matrix mat1, Matrix mat2, Matrix product, int rank, int num_pr
     }
 
     // Gather the local products to the master process
-    MPI_Gatherv(local_product, send_counts[rank], MPI_DOUBLE, product.data, send_counts, displs, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(local_product, send_counts[rank], MPI_FLOAT, product.data, send_counts, displs, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     free(local_mat1);
     free(local_mat2);
@@ -197,9 +197,9 @@ void transpose_mpi(Matrix mat, Matrix trans, int rank, int num_procs) {
     }
 
     // Scatter elements of mat
-    double *local_mat = malloc(send_counts[rank] * sizeof(double));
-    double *local_trans = malloc(send_counts[rank] * sizeof(double));
-    MPI_Scatterv(mat.data, send_counts, displs, MPI_DOUBLE, local_mat, send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    float *local_mat = malloc(send_counts[rank] * sizeof(float));
+    float *local_trans = malloc(send_counts[rank] * sizeof(float));
+    MPI_Scatterv(mat.data, send_counts, displs, MPI_FLOAT, local_mat, send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     // Calculate local transposition
     for (int i = 0; i < send_counts[rank]; i++) {
@@ -210,7 +210,7 @@ void transpose_mpi(Matrix mat, Matrix trans, int rank, int num_procs) {
     }
 
     // Gather the local transpositions to the master process
-    MPI_Gatherv(local_trans, send_counts[rank], MPI_DOUBLE, trans.data, send_counts, displs, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(local_trans, send_counts[rank], MPI_FLOAT, trans.data, send_counts, displs, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     free(local_mat);
     free(local_trans);
@@ -234,9 +234,9 @@ void sum_mpi(Matrix mat, Matrix sum, int rank, int num_procs) {
     }
 
     // Scatter columns of mat
-    double *local_mat = malloc(mat.rows * send_counts[rank] * sizeof(double));
-    double *local_sum = malloc(send_counts[rank] * sizeof(double));
-    MPI_Scatterv(mat.data, send_counts, displs, MPI_DOUBLE, local_mat, mat.rows * send_counts[rank], MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    float *local_mat = malloc(mat.rows * send_counts[rank] * sizeof(float));
+    float *local_sum = malloc(send_counts[rank] * sizeof(float));
+    MPI_Scatterv(mat.data, send_counts, displs, MPI_FLOAT, local_mat, mat.rows * send_counts[rank], MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     // Calculate local column sums
     for (int i = 0; i < send_counts[rank]; i++) {
@@ -244,11 +244,11 @@ void sum_mpi(Matrix mat, Matrix sum, int rank, int num_procs) {
         for (int j = 0; j < mat.rows; j++) {
             colSum += local_mat[j * send_counts[rank] + i];
         }
-        local_sum[i] = colSum;
+        local_sum[i] = (float)colSum;
     }
 
     // Gather the local column sums to the master process
-    MPI_Gatherv(local_sum, send_counts[rank], MPI_DOUBLE, sum.data, send_counts, displs, MPI_DOUBLE, MASTER, MPI_COMM_WORLD);
+    MPI_Gatherv(local_sum, send_counts[rank], MPI_FLOAT, sum.data, send_counts, displs, MPI_FLOAT, MASTER, MPI_COMM_WORLD);
 
     free(local_mat);
     free(local_sum);
@@ -274,8 +274,8 @@ void square_mpi(Matrix mat, int rank, int num_procs) {
     }
 
     // Scatter matrix data
-    double* local_data = malloc(send_counts[rank] * sizeof(double));
-    MPI_Scatterv(mat.data, send_counts, displs, MPI_DOUBLE, local_data, send_counts[rank], MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    float* local_data = malloc(send_counts[rank] * sizeof(float));
+    MPI_Scatterv(mat.data, send_counts, displs, MPI_FLOAT, local_data, send_counts[rank], MPI_FLOAT, 0, MPI_COMM_WORLD);
 
     // Square the local elements
     for (int i = 0; i < send_counts[rank]; i++) {
@@ -283,7 +283,7 @@ void square_mpi(Matrix mat, int rank, int num_procs) {
     }
 
     // Gather squared elements to the root process
-    MPI_Gatherv(local_data, send_counts[rank], MPI_DOUBLE, mat.data, send_counts, displs, MPI_DOUBLE, 0, MPI_COMM_WORLD);
+    MPI_Gatherv(local_data, send_counts[rank], MPI_FLOAT, mat.data, send_counts, displs, MPI_FLOAT, 0, MPI_COMM_WORLD);
 
     free(send_counts);
     free(displs);
